created abstract class Renderer and refactored TF Renderer
